+++ src/lib/types/states.ts
@@ -26,6 +26,7 @@ export interface IScanPageState{
 }
 export interface IUpdatePageState{
      isRequired: boolean,
+     isUpdating: boolean,
      lastUpdated: Date | null,
      log: string[],
      exitMsg: number | null,