+++ src/pages/log-page.tsx
@@ -1,57 +1,17 @@
 import { AppLayout } from "@/components/layout";
-import LogText from "@/components/log";
-import { Button } from "@/components/ui/button";
-import useSettings from "@/hooks/use-settings";
-import { invoke } from "@tauri-apps/api/core";
-import { ChevronLeft, ScrollText } from "lucide-react";
-import { useEffect, useState, useTransition } from "react";
-import { Link, useParams, useSearchParams } from "react-router";
+import LogLoader from "@/loaders/log";
+import { Suspense, lazy} from "react";
+const LogContent = lazy(()=>import("@/contents/log"))
 
 interface Props{
      returnUrl: string
 }
 export default function LogPage({returnUrl}: Props){
-     const {settings} = useSettings()
-     const {logId} = useParams<{ logId: string }>();
-     const [logs, setLogs] = useState<string[]>([]);
-     const [isLoading, startTransition] = useTransition();
-     const [searchParams] = useSearchParams();
-     const category = searchParams.get("category");
-     useEffect(()=>{
-          startTransition(async()=>{
-               if(!logId || !category){
-                    setLogs([`[ERROR] Failed to load the log (Log ID: ${logId})`])
-                    return;
-               }
-               try{
-                    const logs = await invoke<string>("read_log",{
-                         id: logId,
-                         category
-                    })
-                    setLogs(logs.split("\n").filter(Boolean))
-               } catch (err) {
-                    setLogs([`[ERROR] Failed to load the log (Log ID: ${logId})`])
-                    console.error(err)
-               }
-          })
-     },[])
      return (
           <AppLayout className="space-y-4 p-4">
-               <div className="space-y-4">
-                    <h1 className="text-2xl md:text-3xl font-medium border-b pb-2 w-fit">Log Viewer</h1>
-                    {settings.developerMode && (
-                         <p className="text-muted-foreground flex items-center gap-2"><ScrollText className="size-5"/> Log ID: {logId}</p>
-                    )}
-                    <Button asChild size="sm" variant="outline">
-                         <Link to={returnUrl}>
-                              <ChevronLeft/> Back
-                         </Link>
-                    </Button>
-                    <LogText
-                         logs={logs}
-                         isLoading={isLoading}
-                    />
-               </div>
+               <Suspense fallback={<LogLoader/>}>
+                    <LogContent returnUrl={returnUrl}/>
+               </Suspense>
           </AppLayout>
      )
 }